Added check for extension
Received files don't get deleted if there is an exception in
turn_file().  But if it succeeds, then the input file and the
turned one gets deleted.

Not only that but also, I save the input file to disc which
will overwrite whatever is there.  In that case, someone could
send me a file called server.jpg and that would delete overwrite
my image.

Now the file must end with .sgf.  And since there are no SGF files
that I want to keep, this is pretty good.

However, it it fails for a file, then I want to keep that file
so I might save that file as .sgf.failed and refuse to accept
files that failed.

I also need to impose a size limit.  It can be very generous, like
5Mb just to deal with attacks.

+++ app/views.py
@@ -30,16 +30,22 @@ def sgf_turner():
     if request.method == 'POST':
 
         if 'sgf-file' not in request.files:
-            print(''''file' not in request.files''')
+            print('''>>>'file' not in request.files''')
             return redirect(request.url)
 
         sgf_file = request.files['sgf-file']
         sgf_filename = secure_filename(sgf_file.filename)
+
+        if not sgf_filename.endswith('.sgf'):
+            print('''>>> Unaccepted extension for ''' + sgf_filename)
+            return redirect(request.url)
+
+        print('>>> Got sgf file ' + sgf_filename)
         turned_sgf_filename = '.'.join(sgf_filename.split('.')[:-1]) + '_turned.sgf'
         sgf_file.save(sgf_filename)
         turn_file(sgf_filename, turned_sgf_filename)
 
-        print("SGF filename : " + turned_sgf_filename)
+        print(">>> Returning turned sgf file : " + turned_sgf_filename)
         os.remove(sgf_filename)
         ret = send_from_directory('..', turned_sgf_filename, as_attachment=True)
         os.remove(turned_sgf_filename)
